Handle clones/forks of tracked processes (#9)
This PR adds the ability to track processes that were created by a fork/clone.
We send an event to user space when we detect a new PID being created, and the parent PID is already being tracked.

When BTF is enabled we can easily verify in eBPF that the created PID is a process and not a thread - and if it is a thread - we won't bother the user space with that event. When we have BTF we are using a raw tracepoint - which means we can read the task struct of the parent and the child.
When BTF is not enabled, we can't be sure at the clone point if it is a thread or process is created (but we do make sure the parent is tracked before we send an event to user space) - for these cases we verify whether it is a thread or process in user space. When we don't have BTF we use a regular tracepoint and we only have the PIDs of the parent and the child.