Open notifications (#15)
This PR adds a process event of type file_open.
Clients can configure a list of paths which refer to relevant files - if a tracked process would open one of these files, a notification will be sent.
This is done by an additional tracepoint on the openat syscall. The tracepoint program won't get loaded if the user didn't specify any path.

As a result, the events being reported to the client now have 4 possible types - exec, fork, file_open and exit.
Follow up PRs include:

Moving the exe path filter to eBPF.
Improving the duration filter.