Increase the number of environment variables scanned in execve tracepoint (#24)
The goal of this PR is to provide a hot-fix for increasing the amount of environment variable the eBPF code can scan. It is increased from 128 to 1024.
This is possible by 2 changes:

Reducing the max env prefix length from 64 to 16 - it still satisfies any practical use of this feature and allows to reduce the loop complexity for the verifier.
Using an optimized compare function that does a compare on a constant size. This size is currently 16 - and the env prefix can be configured to be shorter than that and the buffers being compared will be padded with zero bytes in that case.
As future improvemnts:

Report to user space when we encounter a process that has more env vars than what we can support.