[RUN-104] add support for older kernels add tests for 4.19 (#32)
Adjustments to support kernel 4.19:

Replace const volatile with a config map that is marked read-only when this kind of flag is supported (from kernel 5.2).
Remove bpf_printk since the strings used end up in the .rodata section - and the .rodata section is mapped to a read-only map - which might not be supported. As a future task a comprehensive mechanism to pass errors/warn events from eBPF code needs to be added.
kernel 4.19 has a limitation of 4096 instructions per program and it does not support bounded loops - which were added in 5.3. Hence all the loops being used in that code needs to have pragam unroll. Doing that increases the instructions count - which get to the 4096 limit. To handle this added a flavor of small program to reduce the loops size - this is mainly relevant for the env vars scan which was adjusted to have 128 max iterations on kernels below 5.3. In order to further decrease the instruction count on older kernels the small program variant does not examine the executable name in the execve program. This can lead to some extra non necessary events being reported on those older kernels - the filtered executable test case is updated to being skipped on older kernels.