feat: support mounted /proc fs to avoid requirement to hostPid namespace (#33)
This PR aims to reduce the permissions, mounts and host namespaces sharing of the detector.

Adding a detector option to allow users to configure the location of the proc filesystem. This is done to remove the requirement for setting hostPID: true.
The example CLI and daemonset are updated to reflect the change.