if we filed to read the root pid ns, fallback to reporting PIDs in that ns and avoid returning an error (#38)
if we failed to read the root pid ns, fallback to reporting PIDs in that ns and avoid returning an error.
Cases where the pid ns is crucial is when the detector is running inside a container without access to the host pid ns (kind cluster on macOS is an example).
When running in k8s, as long as the host /proc dir is mounted the detector can function properly even without knowing the pid ns.