RUN-00 | fix: add open() and openat2() tracepoints for file open tracking (#41)
Make the file open tracking more robust:

Add tracepoints for open and openat2 system call in addition to the current openat one.
Add exit tracepoints to these system calls as well - which is more accurate and allows us to report an event on only successful open operations.
open is the legacy system call - and is not present on arm64 architecture. openat2 is only supported from kernel 5.6 - hence both have conditional logic that they might not be relevant depending on the machine and kernel.
An example case where the current approach does not work: on alpine based images where musl-libc is used - it will fallback to use the open system call on architectures that support it - such as amd64.