RUN-00 | test: add compatibility tests and C written program for testing (#42)
Extending the testing to cover arm64 and alpine variants in addition to the current debian and x86 variants.
Features like the files open tracking are affected by the distribution and architecture combinations (for example alpine based distrubtions might use open system call instead of openat if it is supported by the arch (open is present on x86 but not on arm64) - This was fixed in #41.
In addition adding musl/glibc linked program written in C that is used in the tests. This is added to capture the different behaviors between the distributions.